Fuck it yolo
Most of the things should at least be reasonably safe atm. I'm unable to check if the `cache`/`cached` functions
are properly parameterizing the SQL queries to the `db`-variable as I can't find any details on the implementation,
but from a cursory glance it looks like it at least takes the user data as parameters separated from the query itself.